requirements-all.txt
# Dependencias completas del proyecto ETL Dashboard

# Flask y CORS
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# Base de datos
mysql-connector-python==8.2.0
PyMySQL==1.1.0
SQLAlchemy==2.0.23

# Análisis y predicción
numpy==1.26.2
pandas==2.1.4
scipy==1.11.4

# Utilidades
python-dotenv==1.0.0
python-dateutil==2.8.2

# Servidor de producción
gunicorn==21.2.0
